Various cleanups
- wim_timestamp_to_str() instead of asctime(gmtime(wim_timestamp_to_unix()))
- inode_stream_lte() non-inline to save space
- new errors WIMLIB_ERR_CHAR_CONVERSION and WIMLIB_ERR_INVALID_PART_NUMBER
- write_xml_data():  Have libxml2 write directly to output file with UTF-16LE
  encoding (no need for in-memory buffer and converting to UTF-16LE afterwards)
- read_header(): Validate total_parts and part_number
- Fix a couple potential integer overflows
- Fix some indentation and phase out more 'uint' uses
- Remove some unneeded DEBUG() statements

+++ src/xml.h
@@ -6,7 +6,7 @@
 /* A struct wim_info structure corresponds to the entire XML data for a WIM file. */
 struct wim_info {
 	u64 total_bytes;
-	u64 num_images;
+	int num_images;
 	/* Array of `struct image_info's, one for each image in the WIM that is
 	 * mentioned in the XML data. */
 	struct image_info *images;
